refactor: split the general profile page into my profile and user profile
