+++ src/pages/index.ts
@@ -1,5 +1,6 @@
 export { TodosPage } from './todos-page/todos-page';
 export { ProfilePage } from './profile-page/profile-page';
+export { UserProfilePage } from './user-page/user-page';
 
 export { RegistrationPage } from './auth-pages/registration-page/registration-page';
 export { LoginPage } from './auth-pages/login-page/login-page';
