+++ src/pages/user-page/user-page.module.scss
@@ -1,0 +1,14 @@
+.card {
+  max-width: 650px;
+  width: 100%;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+}
+
+.descriptors {
+  margin-bottom: 24px;
+}
+
+.buttons-wrapper {
+  display: flex;
+  gap: 5px;
+}
